requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Async Task Queue
celery[redis]==5.3.4
redis==4.6.0
redis[hiredis]==4.6.0
fakeredis==2.32.0
flower==2.0.1

# High Performance
aiofiles==23.2.1
aiohttp==3.9.1
uvloop==0.19.0

# Machine Learning & OCR
torch==2.1.0
transformers==4.30.0
tokenizers==0.13.3
sentencepiece==0.2.1
easyocr==1.7.0
accelerate==0.20.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==9.5.0

# HTML Processing
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Monitoring & Logging
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0
httpx==0.25.1
faker==20.1.0
locust==2.18.0

# HTTP Client (for health checks)
requests==2.31.0

# Additional dependencies for image processing and file handling
pathlib2==2.3.7
